MIM-95 split workspace presentation helpers
